Fix all C++ compilation warnings in gcc 12.2 and clang 15 (#179)
* Silence compiler warnings coming from rapidjson

We have no control over the code in that header and we do not
want its warnings to pollute our build.

* Silence warnings about control reaching end of non-void function

Let's make it clear to the compiler that we expect the switch-case
to be exhaustive. This has the additional advantage of noisily
breaking in case in the future this is not the case anymore.

* Fix uninitialized variable warning in correction.cc

* Fix warning about unsigned/signed comparison in python.cc

---------

+++ src/python.cc
@@ -50,7 +50,7 @@ namespace {
     double * outptr = static_cast<double*>(outbuffer.ptr);
     {
       py::gil_scoped_release release;
-      for (size_t i=0; i < outbuffer.shape[0]; ++i) {
+      for (long i=0; i < outbuffer.shape[0]; ++i) {
         for (const auto& varg : vargs) {
           if ( std::holds_alternative<int>(inputs[varg.first]) ) {
             inputs[varg.first] = static_cast<int*>(varg.second.ptr)[i];
